User Check for editing and deleting blogs

+++ odyssey_django/blogs/views.py
@@ -37,19 +37,24 @@ class BlogDetail(APIView):
 
     def put(self, request , id):
         blog = self.get_object(id)
-        place = Place.objects.get(id = request.data["place"])
-        blog.place = place
-        blog.save()
-        serializer = BlogSerializer(blog, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        current_user = Traveller.objects.get(username = request.user)
+        if (current_user==blog.author):
+            place = Place.objects.get(id = request.data["place"])
+            blog.place = place
+            blog.save()
+            serializer = BlogSerializer(blog, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
         blog = self.get_object(id)
-        blog.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        current_user = Traveller.objects.get(username = request.user)
+        if (current_user==blog.author):
+            blog.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
 class AddBlog(APIView):
     def post(self, request):
